工具函数整理
