--- interface_struct/server_interface.go
@@ -1,7 +1,0 @@
-package interface_struct
-
-type ServerInterface interface {
-	Listen(port string)
-	Loop()
-	Close()
-}
